scripts/requirements.txt
minio==7.2.0
langfuse>=2.43.0
python-dotenv>=1.0.0